Move the rendezvous session handler to the _synapse/ namespace
